Created Wallet service Moved wallet source code
