+++ serverless.yml
@@ -37,71 +37,6 @@ package:
   individually: true
 
 functions:
-  wallet:
-    handler: src/lambdas/wallet/walletLambda.handle
-    events:
-      - http:
-          path: account/{accountId}/wallet
-          method: get
-          authorizer:
-            arn: ${self:custom.cognitoIdpArn}
-            scopes:
-              - email
-          parameters:
-            paths:
-              accountId: true
-      - http:
-          path: account/{accountId}/wallet
-          method: POST
-          authorizer:
-            arn: ${self:custom.cognitoIdpArn}
-            scopes:
-              - email
-          parameters:
-            paths:
-              accountId: true
-      - http:
-          path: account/{accountId}/wallet/{walletId}
-          method: GET
-          authorizer:
-            arn: ${self:custom.cognitoIdpArn}
-            scopes:
-              - email
-          parameters:
-            paths:
-              accountId: true
-              walletId: true
-      - http:
-          path: account/{accountId}/wallet/{walletId}
-          method: PUT
-          authorizer:
-            arn: ${self:custom.cognitoIdpArn}
-            scopes:
-              - email
-          parameters:
-            paths:
-              accountId: true
-              walletId: true
-      - http:
-          path: account/{accountId}/wallet/{walletId}
-          method: DELETE
-          authorizer:
-            arn: ${self:custom.cognitoIdpArn}
-            scopes:
-              - email
-          parameters:
-            paths:
-              accountId: true
-              walletId: true
-    package:
-      include:
-        - src/lambdas/wallet/**
-        - libs/*
-        - node_modules/**
-      exclude:
-        - test/**
-        - .git/**
-        - .travis.yml
   requestFileUpload:
     handler: src/lambdas/request-file-upload/requestFileUploadLambda.handle
     events:
