+++ services/wallet/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "virtwallet-wallet",
+  "version": "1.0.0",
+  "description": "Virtwallet Wallet Service",
+  "repository": "https://github.com/silvamb/virtwallet-serverless.git",
+  "author": "Matheus Silva <crush.comp04@gmail.com>",
+  "license": "GPLv3",
+  "dependencies": {
+    "libs": "file:../../libs"
+  },
+  "devDependencies": {
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
+    "mocha": "^7.0.0",
+    "serverless-iam-roles-per-function": "^2.0.2"
+  },
+  "scripts": {
+    "test": "mocha ./test/*-test.js"
+  }
+}
